requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
croniter==1.3.14
dateutils==0.6.12
deepdiff==6.3.0
docker-pycreds==0.4.0
fastapi==0.88.0
filelock==3.12.0
flake8==6.0.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
inquirer==3.1.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.1
pathtools==0.1.2
platformdirs==3.5.0
protobuf==4.23.0
psutil==5.9.5
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.29.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.22.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.0
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
wandb==0.15.2
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2